+++ cluster/pulumi/canton-network/src/sv.ts
@@ -163,8 +163,7 @@ export function installSvNode(
     values.joinWithKeyOnboarding = {
       publicKey: joinWithKey.publicKey,
       privateKey: joinWithKey.privateKey,
-      sponsorApiUrl: 'http://sv-app.sv-1',
-      sponsorApiPort: 5014,
+      sponsorApiUrl: 'http://sv-app.sv-1:5014',
       svcApiAddress: 'svc-app.svc',
     };
   }