+++ cluster/images/docs/src/release_notes.rst
@@ -12,6 +12,10 @@ Release Notes
     and similarly, the underlying image has been renamed to ``external-proxy-sv``. They will soon be
     simplified to include only the configuration essential for an SV node deployment.
   * ``joinWithKeyOnboarding.keyName`` in ``sv-values.yaml`` has been renamed to ``onboardingName``.
+  * ``svSponsorPort`` in ``validator-values.yaml`` has been removed. The port is now included in ``svSponsorAddress``. The default sponsor address has
+    been changed to ``https://sv.sv-1.svc.TARGET_CLUSTER.network.canton.global/api/v0/sv``.
+  * ``sponsorApiPort`` in ``sv-values.yaml`` has been removed. The port is now included in ``sponsorApiUrl``. The default sponsor address has
+    been changed to ``https://sv.sv-1.svc.TARGET_CLUSTER.network.canton.global/api/v0/sv``.
 
 2023-05-21
 ----------